backend/app/utils/image_processing.py
"""Image processing utilities"""

from PIL import Image
import io
import logging
from typing import Optional

logger = logging.getLogger(__name__)


class ImageProcessor:
    """Utility for processing images"""
    
    async def extract_text(self, image_bytes: bytes) -> str:
        """
        Extract text from image using OCR
        
        Args:
            image_bytes: Image file bytes
            
        Returns:
            Extracted text
        """
        try:
            logger.info("Extracting text from image")
            
            # Open image
            image = Image.open(io.BytesIO(image_bytes))
            
            # TODO: Implement OCR using pytesseract or similar
            # For now, return placeholder
            extracted_text = "Extracted text placeholder"
            
            return extracted_text
            
        except Exception as e:
            logger.error(f"Error extracting text from image: {e}", exc_info=True)
            raise
    
    async def resize_image(
        self,
        image_bytes: bytes,
        width: int,
        height: int
    ) -> bytes:
        """Resize image"""
        try:
            image = Image.open(io.BytesIO(image_bytes))
            resized = image.resize((width, height))
            
            output = io.BytesIO()
            resized.save(output, format=image.format)
            return output.getvalue()
            
        except Exception as e:
            logger.error(f"Error resizing image: {e}", exc_info=True)
            raise
    
    async def optimize_image(self, image_bytes: bytes, quality: int = 85) -> bytes:
        """Optimize image size and quality"""
        try:
            image = Image.open(io.BytesIO(image_bytes))
            
            output = io.BytesIO()
            image.save(output, format=image.format, optimize=True, quality=quality)
            return output.getvalue()
            
        except Exception as e:
            logger.error(f"Error optimizing image: {e}", exc_info=True)
            raise

