backend/requirements.txt
# FastAPI and dependencies
fastapi[all]==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# AI/ML
anthropic==0.21.3

# Database
supabase==2.4.0
psycopg2-binary==2.9.9

# Document Processing
python-docx==1.1.0
reportlab==4.1.0
Pillow==10.2.0
pypdf2==3.0.1

# Settings and Configuration
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
aiofiles==23.2.1

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.27.0

